AG-5382-fixes2
